Add position model,add positon handler, implement logics related to openning and closeing positions + minore changes

+++ internal/model/base.go
@@ -7,7 +7,7 @@ import (
 )
 
 type BaseEntity struct {
-	ID        int            `json:"id" gorm:"primaryKey"`
+	ID        int64          `json:"id" gorm:"primaryKey"`
 	CreatedAt time.Time      `json:"createdAt"`
 	UpdatedAt time.Time      `json:"updatedAt"`
 	DeletedAt gorm.DeletedAt `json:"-" gorm:"index"`
